submit_2

+++ srcs/draw_line_utils.c
@@ -6,7 +6,7 @@
 /*   By: hyuki <hyuki@student.42tokyo.jp>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/12/05 12:09:25 by hyuki             #+#    #+#             */
-/*   Updated: 2020/12/05 16:42:10 by hyuki            ###   ########.fr       */
+/*   Updated: 2020/12/08 13:33:03 by hyuki            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -51,7 +51,7 @@ int		ft_draw_line_calc(t_s *s)
 		s->g.d_c.texx = floor(s->g.d_c.wallx * s->g.ws.now.width);
 	else
 		s->g.d_c.texx = floor((1 - s->g.d_c.wallx) * s->g.ws.now.width);
-	s->g.d_c.step = (double)s->g.ws.now.height / s->g.d_c.lineh;
+	s->g.d_c.step = (double)s->g.ws.now.height / (double)s->g.d_c.lineh;
 	if (s->g.d_c.lineh > s->g.img.scre_height)
 		s->g.d_c.texpos = s->g.ws.now.height *
 			((1 - (double)s->g.img.scre_height / (double)s->g.d_c.lineh) / 2);
